Permiso borrar almacenes

+++ php/control_almacen.php
@@ -106,24 +106,32 @@ switch ($Accion) {
         // $Accion es igual a 3 realiza:
         //CON POST RECIBIMOS LA VARIABLE DEL BOTON POR EL SCRIPT DE "almacenes_punto_venta.php" QUE NESECITAMOS PARA BORRAR
         $id = $conn->real_escape_string($_POST['id']);
-    	 #SELECCIONAMOS LA INFORMACION A BORRAR
-        $almacen = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM `punto_venta_almacenes` WHERE id = $id"));
-        #CREAMOS EL SQL DE LA INSERCION A LA TABLA  `pv_borrar_almacenes` PARA NO PERDER INFORMACION
-        $sql = "INSERT INTO `pv_borrar_almacenes` (id_almacen, nombre, registro, borro, fecha_borro) 
-                VALUES($id, '".$almacen['nombre']."', '".$almacen['usuario']."', '$id_user','$Fecha_hoy')";
-        //VERIFICAMOS QUE LA SENTECIA FUE EJECUTADA CON EXITO!
-        if(mysqli_query($conn, $sql)){
-            //SI DE CREA LA INSERCION PROCEDEMOS A BORRRAR DE LA TABLA `punto_venta_almacenes`
-            #VERIFICAMOS QUE SE BORRE CORRECTAMENTE EL CLIENTE DE `punto_venta_almacenes`
-            if(mysqli_query($conn, "DELETE FROM `punto_venta_almacenes` WHERE `punto_venta_almacenes`.`id` = $id")){
-            #SI ES ELIMINADO MANDAR MSJ CON ALERTA
-                echo '<script >M.toast({html:"Almacen borrado con exito.", classes: "rounded"})</script>';
-                echo '<script>recargar_almacen_lista()</script>';// REDIRECCIONAMOS (FUNCION ESTA EN ARCHIVO modals.php)
-            }else{
-            #SI NO ES BORRADO MANDAR UN MSJ CON ALERTA
-                echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});/script>";
-            }
-        } 
+    	//Obtenemos la informacion del Usuario
+        $User = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM `users` WHERE user_id = $id_user"));
+        //SE VERIFICA SI EL USUARIO LOGEADO TIENE PERMISO DE BORRAR ARTICULOS
+        if ($User['b_almacenes'] == 1) {
+            #SELECCIONAMOS LA INFORMACION A BORRAR
+            $almacen = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM `punto_venta_almacenes` WHERE id = $id"));
+            #CREAMOS EL SQL DE LA INSERCION A LA TABLA  `pv_borrar_almacenes` PARA NO PERDER INFORMACION
+            $sql = "INSERT INTO `pv_borrar_almacenes` (id_almacen, nombre, registro, borro, fecha_borro) 
+                    VALUES($id, '".$almacen['nombre']."', '".$almacen['usuario']."', '$id_user','$Fecha_hoy')";
+            //VERIFICAMOS QUE LA SENTECIA FUE EJECUTADA CON EXITO!
+            if(mysqli_query($conn, $sql)){
+                //SI DE CREA LA INSERCION PROCEDEMOS A BORRRAR DE LA TABLA `punto_venta_almacenes`
+                #VERIFICAMOS QUE SE BORRE CORRECTAMENTE EL CLIENTE DE `punto_venta_almacenes`
+                if(mysqli_query($conn, "DELETE FROM `punto_venta_almacenes` WHERE `punto_venta_almacenes`.`id` = $id")){
+                #SI ES ELIMINADO MANDAR MSJ CON ALERTA
+                    echo '<script >M.toast({html:"Almacen borrado con exito.", classes: "rounded"})</script>';
+                    echo '<script>recargar_almacen_lista()</script>';// REDIRECCIONAMOS (FUNCION ESTA EN ARCHIVO modals.php)
+                }else{
+                #SI NO ES BORRADO MANDAR UN MSJ CON ALERTA
+                    echo "<script >M.toast({html: 'Ha ocurrido un error.', classes: 'rounded'});/script>";
+                }
+            } 
+        }else{
+            echo '<script >M.toast({html:"Permiso denegado.", classes: "rounded"});
+            M.toast({html:"Comunicate con un administrador.", classes: "rounded"});</script>';
+        }   
         break;
 }// FIN switch
 mysqli_close($conn);
